% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-advance-decline.R
\name{nse_advances_declines}
\alias{nse_advances_declines}
\title{Advances & Declines}
\usage{
nse_advances_declines(clean_names = TRUE)
}
\arguments{
\item{clean_names}{Logical; if \code{TRUE}, makes the column names
descriptive and uses snake_case.}
}
\value{
A tibble with the following columns:

\item{index_name}{Name of the index.}
\item{advances}{Number of stocks increased (in green).}
\item{declines}{Number of stocks decreased (in red).}
\item{unchanged}{Number of stocks unchanged.}
}
\description{
NSE indices advances & declines.
}
\examples{
\donttest{
nse_advances_declines()

# retain original column names as returned by NSE
nse_advances_declines(clean_names = FALSE)
}

}
