% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-fo.R
\name{nse_fo_top_losers}
\alias{nse_fo_top_losers}
\title{NSE top F & O losers}
\usage{
nse_fo_top_losers()
}
\description{
Top futures and options losers for the last trading session.
}
\examples{
\dontrun{
nse_fo_top_losers()
}

}
