% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-index.R
\name{nse_index_list}
\alias{nse_index_list}
\title{Index list}
\usage{
nse_index_list(clean_names = TRUE)
}
\arguments{
\item{clean_names}{Logical; if \code{TRUE}, makes the column names
descriptive and uses snake_case.}
}
\value{
A tibble with the following column:

\item{index_name}{Name of the index.}
}
\description{
List NSE indices.
}
\examples{
\dontrun{
nse_index_list()

# retain original column names as returned by NSE
nse_index_list(clean_names = FALSE)
}

}
