% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-index.R
\name{nse_index_quote}
\alias{nse_index_quote}
\title{Index quote}
\usage{
nse_index_quote(clean_names = TRUE)
}
\arguments{
\item{clean_names}{Logical; if \code{TRUE}, makes the column names
descriptive and uses snake_case.}
}
\value{
A tibble with the following columns:

\item{index_name}{Name of the NSE indices.}
\item{last_traded_price}{Last traded price.}
\item{change}{Change in price.}
\item{percent_change}{Percentage change in price.}
}
\description{
Fetch the quote for a given index.
}
\examples{
\dontrun{
nse_index_quote()

# retain original column names as returned by NSE
nse_index_quote(clean_names = FALSE)
}

}
