% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-preopen.R
\name{nse_preopen_nifty_bank}
\alias{nse_preopen_nifty_bank}
\title{Pre open nifty bank}
\usage{
nse_preopen_nifty_bank()
}
\description{
Fetch data of pre open session of Nifty Bank from NSE.
}
\examples{
\donttest{
nse_preopen_nifty_bank()
}

}
