% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_code}
\alias{nse_stock_code}
\title{Stock code}
\usage{
nse_stock_code(clean_names = TRUE)
}
\arguments{
\item{clean_names}{Logical; if \code{TRUE}, makes the column names
descriptive and uses snake_case.}
}
\value{
A tibble with the following columns:

\item{symbol}{NSE ticker.}
\item{company}{Name of the firm.}
}
\description{
Fetch stock symbol and name from NSE.
}
\examples{
\dontrun{
nse_stock_code()

# retain original column names as returned by NSE
nse_stock_code(clean_names = FALSE)
}

}
