% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_top_gainers}
\alias{nse_stock_top_gainers}
\title{NSE top gainers}
\usage{
nse_stock_top_gainers()
}
\description{
Top gainers for the last trading session.
}
\examples{
\dontrun{
nse_stock_top_gainers()
}

}
