% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_top_losers}
\alias{nse_stock_top_losers}
\title{NSE top losers}
\usage{
nse_stock_top_losers()
}
\description{
Fetch top losers for the last trading session.
}
\examples{
nse_stock_top_losers()

}
