% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_year_low}
\alias{nse_stock_year_low}
\title{52 week low}
\usage{
nse_stock_year_low()
}
\description{
Stocks that have touched their 52 week lows during the day on NSE.
}
\examples{
\dontrun{
nse_stock_year_low()
}

}
