% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-preopen.R
\name{preopen_nifty}
\alias{preopen_nifty}
\alias{nse_preopen_nifty}
\alias{nse_preopen_nifty_bank}
\title{Pre open Nifty data}
\usage{
nse_preopen_nifty()

nse_preopen_nifty_bank()
}
\value{
A tibble with the following columns:

\code{symbol}{NSE ticker.}
\code{series}{Equity (EQ)}
\code{corp_action_date}{Corporate action date.}
\code{corp_action}{Corporate Action}
\code{price}{Price}
\code{change}{Change in price}
\code{percent_change}{Percentage change in price.}
\code{prev_close}{Previous close.}
\code{quantity}{Quantity}
\code{value}{Value (in lakhs),}
\code{mkt_cap}{Free float market capitalization (in crores).}
\code{year_high}{Normal market 52 week high.}
\code{year_low}{Normal market 52 week low.}
}
\description{
Fetch data of pre open session of Nifty & Nifty Bank.
}
\examples{
\donttest{
nse_preopen_nifty()
nse_preopen_nifty_bank()
}

}
