% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-advance-decline.R
\name{nse_advances_declines}
\alias{nse_advances_declines}
\title{Advance declines}
\usage{
nse_advances_declines()
}
\value{
A tibble with the following columns:

\code{index_name}{Name of the index.}
\code{advances}{Number of stocks increased (in green).}
\code{declines}{Number of stocks decreased (in red).}
\code{unchanged}{Number of stocks unchanged.}
}
\description{
NSE indices advances & declines.
}
\examples{
\dontrun{
nse_advances_declines()
}

}
