% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-index.R
\name{nse_index_list}
\alias{nse_index_list}
\title{Index list}
\usage{
nse_index_list()
}
\value{
A tibble with the following column:

\item{index_name}{Name of the index.}
}
\description{
List NSE indices.
}
\examples{
\donttest{
nse_index_list()
}

}
