% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-index.R
\name{nse_index_valid}
\alias{nse_index_valid}
\title{Validate index symbol.}
\usage{
nse_index_valid(index_code)
}
\arguments{
\item{index_code}{Symbol of the index.}
}
\description{
Check if index symbol is valid.
}
\examples{
\donttest{
nse_index_valid("nifty auto")
nse_index_valid("nifty cps")
}

}
