% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_code}
\alias{nse_stock_code}
\title{Stock code}
\usage{
nse_stock_code()
}
\value{
A tibble with the following columns:

\item{symbol}{NSE ticker.}
\item{company}{Name of the firm.}
}
\description{
Fetch stock symbol and name from NSE.
}
\examples{
\donttest{
nse_stock_code()
}

}
