% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_most_traded}
\alias{nse_stock_most_traded}
\title{Most actively traded stocks}
\usage{
nse_stock_most_traded()
}
\description{
Fetch most actively traded stocks in a month on NSE.
}
\examples{
\dontrun{
nse_stock_most_traded()
}

}
