% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_quote}
\alias{nse_stock_quote}
\title{Stock quote}
\usage{
nse_stock_quote(stock_code)
}
\arguments{
\item{stock_code}{Symbol of the stock.}
}
\description{
Fetch the quote for a given stock code.
}
\examples{
nse_stock_quote("infy")

}
