% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_valid}
\alias{nse_stock_valid}
\title{Valid stock code}
\usage{
nse_stock_valid(stock_code)
}
\arguments{
\item{stock_code}{Symbol of the stock.}
}
\description{
Check if stock code is valid.
}
\examples{
\donttest{
nse_stock_valid("infy")
nse_stock_valid("glo")
}

}
