% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-stock.R
\name{nse_stock_year_high}
\alias{nse_stock_year_high}
\title{52 week high}
\usage{
nse_stock_year_high()
}
\description{
Stocks that have touched their 52 week highs during the day on NSE.
}
\examples{
\dontrun{
nse_stock_year_high()
}

}
